Switched to a simpler XP scheme
Like in OD&D, XP is now simply HD×100. This is simpler and doesn't
require the referee to wonder for every special ability whether it is
worthy of an asteriks and thus of the extra xp or not. This makes it
simpler to determine XP because you also don't need the table at hand.
Also, I haven't heard of anybody's campaign getting unbalanced because
of it.